+++ services/gundeck/src/Main.hs
@@ -4,12 +4,11 @@ import Imports
 import Gundeck.API
 import OpenSSL (withOpenSSL)
 
-import Gundeck.Options
 import Util.Options
 
 main :: IO ()
 main = withOpenSSL $ do
-    options <- getOptions desc (Just optsParser) defaultPath
+    options <- getOptions desc Nothing defaultPath
     runServer options
   where
     desc = "Gundeck - Push Notification Hub Service"